Split files properly
