+++ packages/alfa-dom/src/node/element/predicate/has-input-type.ts
@@ -2,72 +2,11 @@ import { Predicate } from "@siteimprove/alfa-predicate";
 import { Refinement } from "@siteimprove/alfa-refinement";
 
 import { Element } from "../../element";
+import { InputType, inputType } from "../input-type";
 import { hasName } from "./has-name";
 
 const { equals, test } = Predicate;
 const { and } = Refinement;
-
-/**
- * {@link https://html.spec.whatwg.org/#attr-input-type}
- */
-type InputType =
-  | "hidden"
-  | "search"
-  | "tel"
-  | "url"
-  | "email"
-  | "password"
-  | "date"
-  | "month"
-  | "week"
-  | "time"
-  | "datetime-local"
-  | "number"
-  | "range"
-  | "color"
-  | "checkbox"
-  | "radio"
-  | "file"
-  | "submit"
-  | "image"
-  | "reset"
-  | "button"
-  | "text";
-
-/**
- * @public
- */
-export function inputType(element: Element<"input">): InputType {
-  return element
-    .attribute("type")
-    .flatMap((attribute) =>
-      attribute.enumerate(
-        "hidden",
-        "search",
-        "tel",
-        "url",
-        "email",
-        "password",
-        "date",
-        "month",
-        "week",
-        "time",
-        "datetime-local",
-        "number",
-        "range",
-        "color",
-        "checkbox",
-        "radio",
-        "file",
-        "submit",
-        "image",
-        "reset",
-        "button",
-        "text"
-      )
-    )
-    .getOr("text");
-}
 
 /**
  * @public
