+++ packages/alfa-dom/tsconfig.json
@@ -23,6 +23,7 @@
     "src/node/comment.ts",
     "src/node/document.ts",
     "src/node/element.ts",
+    "src/node/element/input-type.ts",
     "src/node/element/predicate.ts",
     "src/node/element/predicate/has-attribute.ts",
     "src/node/element/predicate/has-display-size.ts",
